Implements create a wallet
First draft of the core domain.

Add the first primary port, which shapes the contract how to talk to the
domain.

Create `WalletId` micro type and `Wallet` type.

Currently, the in-memory implementation of the port holds the wallets
created in a map, so that we don't delay writing tests for the contract.