Implement http adapter
We are now creating an http adapter that communicates with the domain
port to exercise the logic.

We imported the domain project as well as the domain port test contract
testFixture, so we can verify the infrastructure we used, http, can
comply with the domain contract.

We started off by implementing the web controller, which uses the port.
Then setup our test suite, referencing the contract we took from the
testFixtures. To pass the contract test, we create an http client that
implements the domain port, calling the web controller that adapts to
the domain port.