Implements mongodb driven port adapter.
A bit of experiment here...

We attempt to implement a nosql adapter to the driven port of the
domain; our choice for now is MongoDb.

We used a kotlin native library to integrate with MongoDb to make things
a bit easier.

First things first is to try spinning up a docker container to verify we
can use the library and connect to a MongoDb instance. Afer that passed,
we were ready to start our TDD.

We used testcontainers to test our adapter can succesfully connect to a
dockerized MongoDb instance. After creating a `TestContainersBase`
object that via that we start a docker container, we have to make sure
that our driven adapter (technology-specific implementation of our
driven port in domain) complies with the contract we have defined in the
testFixture of the domain.

After some thought of trying to pass the container's randomly generated
port into our a client connecting to it, managed to finally have a green
test. Passing the rest of the tests, and as attempted, required us to
clean the state of the MongoDb docker instance; even though it was
restarting before each test, the state was there. So we implemented a
clean method only for tests, glued this onto our test, after each test
finishes, and we had the rest of the green to satisfy the contract.