Tidy up
Renames in memory driven port to fake.
Wraps in memory driving port into an object to be more readable.
Renames test to in memory.
Renames to just http test because it is a real one.
Refactors test containers base to be a mongo db docker container.
Removes redundant jvm static before each method.