Updates no sql repository dependencies
NoSql repository should depend on a mongodb client. The way this is
being passed in, it is up to a factory's responsibility.

That means now clients that depend on the module responsible for
mongodb, should not depend on its internal usage in production code. But
rather the very class that deals with the business logic.

Only a test dependency has been added to the main application module as
it needs to clean the state in an integration test.