Creates a card wallet http server agnostic of http technology
The main application should not depend on the module's implementation
details about http. It should just ask for an http server without
knowing what kind of http server it is, because this is being taken care
of the very module that knows http.

Note that the testFixtures project still needs to know about http as it
builds an http client that is running against the card wallet http api.